feat: 유저 엔티티 기록 값 변경 로직 수정, API 추가
- 퀘스트 등록수 증가 시 마지막 등록일 갱신
- 퀘스트 완료수 증가 시 마지막 완료일 갱신
- 경험치 골드 획득 시 골드 획득량 갱신
- 골드 획득, 사용량 타입 Long으로 변경
- 업적에서 모든 기록 값을 수용하기 위해 현재 값 타입을 Long으로 변경

+++ dailyquest_web/src/main/java/dailyquest/user/dto/UserPrincipal.kt
@@ -34,8 +34,8 @@ class UserPrincipal(
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
     val lastQuestCompletionDate: LocalDate? = null,
     val perfectDayCount: Int = 0,
-    val goldEarnAmount: Int = 0,
-    val goldUseAmount: Int = 0,
+    val goldEarnAmount: Long = 0,
+    val goldUseAmount: Long = 0,
 ) : UserDetails {
 
     @JsonIgnore
